feat: introduce tycho-protobuf crate
Factors out a new `tycho-protobuf` crate from `protocols/substreams/crates/tycho-substreams`.
The crate lives at `crates/tycho-protobuf` inside the main workspace.

- `src/pb/` holds the generated protobuf types (from `proto/tycho/evm/v1/common.proto`);
  regenerate with `PROTO_BUILD=1 cargo build -p tycho-protobuf` or `buf generate` from the
  crate directory (uses the same `buf.gen.yaml` plugin/attribute config as tycho-substreams).
- `src/convert.rs` exposes the `TryFromMessage` trait and all implementations converting
  protobuf messages to `tycho-common` domain types.
- `src/error.rs` defines `DecodeError` with `Decode(String)` and `Empty` variants so the
  crate has no dependency on the indexer-internal `ExtractionError`.

`BlockChanges` and `TxWithContractChanges` are moved from `tycho-indexer::extractor::models`
to `tycho-common::models::blockchain` and re-exported from there. `into_aggregated()` now
returns `MergeError` instead of `ExtractionError`.

`tycho-indexer` drops the `tycho-substreams` dependency and uses `tycho-protobuf` instead.
`protobuf_deserialisation.rs` is reduced to a single re-export of `TryFromMessage`.
