+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/tycho-storage",
     "crates/tycho-ethereum",
     "crates/tycho-indexer",
+    "crates/tycho-protobuf",
     "crates/tycho-simulation",
     "crates/tycho-test",
     "crates/tycho-integration-test",
@@ -32,6 +33,7 @@ readme = "README.md"
 tycho                  = { path = "crates/tycho",                  version = "0.283.1" }
 tycho-common           = { path = "crates/tycho-common",           version = "0.283.1" }
 tycho-client           = { path = "crates/tycho-client",           version = "0.283.1" }
+tycho-protobuf         = { path = "crates/tycho-protobuf",         version = "0.283.1" }
 tycho-storage          = { path = "crates/tycho-storage",          version = "0.283.1" }
 tycho-ethereum         = { path = "crates/tycho-ethereum",         version = "0.283.1" }
 tycho-indexer          = { path = "crates/tycho-indexer",          version = "0.283.1" }
