+++ crates/tycho-protobuf/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod convert;
+pub mod error;
+pub mod pb;
